Backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML Serving (no training packages needed)
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# File handling
aiofiles==23.2.1

# Optional (if you need them)
scipy==1.11.4  # Only if using scipy in explainable_ai.py
scikit-image==0.22.0  # Only if using skimage functions